freetype: host-freetype-patch depends on host-automake
The "make host-freetype-patch" fails because it doesn't find the
autogen.sh script, the problem was already fixed for "make
freetype-patch" adding the "host-automake" as a prerequisite to
freetype-patch, so fix it by adding "host-automake" as a prerequisite to
"host-freetype-patch" too.

This change also fix the support to top-level parallel make.

Signed-off-by: Fabio Porcedda <fabio.porcedda@gmail.com>