packages: fix and improve support for top-level parallel makefile
To fix packages that fail to build when PARALLEL_JOBS is empty instead
of using an empty PARALLEL_JOBS just avoid to using it in
the MAKE variable.

Check the MAKEFLAGS variable to know automatically if the -j option is
being used, but use the "=" operator instead of the ":=" operator
because it can be checked only in a "recursively expanded variable".
Use "override" because otherwise it's impossible to change the
automatic variable "MAKE".

Signed-off-by: Fabio Porcedda <fabio.porcedda@gmail.com>
Acked-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>