infra: add per-package staging feature
To improve the build reproducibility, instead of sharing a common
staging are, every package has a private staging directory where only
the declared dependency are present.

Having a per-package staging directory ensure that every package use
only the declared dependency, because even if a undeclared dependency is
build before, it will be not included in the private package directory.

This feature is useful for both top-level non-parallel make and
top-level parallel make.

This feature is useful for target dependencies but not for host
dependencies, to solve that a similar feature for host packages must be
developed.

Implementation details:
To minimize change the STAGING_DIR variable will be redefined for each
package to point to the per-package staging directory.

At the beginning of each configuration stage the per-package staging
directory will be created copying using hard links all the declared
target dependencies.
Also each path of every copied configuration files will be changed to
point to the per-package

Signed-off-by: Fabio Porcedda <fabio.porcedda@gmail.com>